Experimental Dash support Experimental BMFF Merkle generation Support for very large assets Start of fast parallel hash work

+++ sdk/src/assertions/box_hash.rs
@@ -39,10 +39,10 @@ pub struct BoxMap {
     pub pad: ByteBuf,
 
     #[serde(skip)]
-    pub range_start: usize,
+    pub range_start: u64,
 
     #[serde(skip)]
-    pub range_len: usize,
+    pub range_len: u64,
 }
 
 /// Helper class to create BoxHash assertion
